refactor: rename commands and method
folderDelete & folderRename
